avoid reading files that are not symlinks when core.symlinks=false
This hack is only needed on FAT filesystems, so there's no point in doing
it the rest of the time. And it's possible for there to be a false
positive, so it's best to avoid the hack when possible.

(cherry picked from commit ae341c1a37eecc1724517e3e025d144badb5abfe)